ItalianRestaurantBackend/src/main/java/com/example/italianrestaurant/meal/MealService.java
package com.example.italianrestaurant.meal;

import jakarta.persistence.EntityNotFoundException;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
@RequiredArgsConstructor
public class MealService {

    private final MealRepository mealRepository;

    public List<Meal> getAllMeals() {
        return mealRepository.findAll();
    }

    public Meal getMealById(Long id) {
        Optional<Meal> meal = mealRepository.findById(id);

        return meal.orElseThrow(EntityNotFoundException::new);
    }
}
